Move drive downloading to spreadsheet manager
