+++ lib/ad_localize.rb
@@ -17,6 +17,7 @@ require 'ad_localize/version'
 require 'ad_localize/ad_logger'
 require 'ad_localize/constant'
 require 'ad_localize/csv_file_manager'
+require 'ad_localize/spreadsheet_manager'
 require 'ad_localize/csv_parser'
 require 'ad_localize/option_handler'
 require 'ad_localize/runner'
